+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0] - June , 2020
+⚠ **Major Update (Breaking Changes)**
+
+### Added
+- Add API key authorizations
+- Custom API URL support
+
+### Changed
+- Due to Judge0 API going [freemium](https://github.com/judge0/api/issues/171), the default API URL, [https://api.judge0.com] is no longer available. To use the API signup for a plan on [RapidAPI](https://rapidapi.com/hermanzdosilovic/api/judge0/pricing) & use the token to access the API through coderunner.
+
+
 ## [0.8] - May 27, 2020
 
 ### Fix